refactor: move metadata field helpers into dedicated modules
