+++ src/views/export/extra-field-list.tsx
@@ -1,5 +1,6 @@
 import { useReconciledBuffer } from "@/hooks/useReconciledBuffer";
 import { Button } from "@/ui/button";
+import { type Pair, pairsToRecord, reconcilePairs, sameRecord, seedPairs } from "@/views/export/extra-field-pairs";
 import { INPUT_STYLES } from "@/views/export/metadata-field-list";
 import { IconPlus, IconX } from "@tabler/icons-react";
 import { nanoid } from "nanoid";
@@ -10,37 +11,6 @@ interface ExtraFieldListProps {
   values: Record<string, string>;
   onChange: (next: Record<string, string>) => void;
 }
-
-interface Pair {
-  id: string;
-  key: string;
-  value: string;
-}
-
-// -- Reconciliation -----------------------------------------------------------
-
-const seedPairs = (extra: Record<string, string>): Pair[] =>
-  Object.entries(extra).map(([key, value]) => ({ id: nanoid(), key, value }));
-
-const reconcilePairs = (previous: Pair[], extra: Record<string, string>): Pair[] =>
-  Object.entries(extra).map(([key, value], index) =>
-    previous[index]?.key === key && previous[index]?.value === value
-      ? previous[index]
-      : { id: previous[index]?.id ?? nanoid(), key, value },
-  );
-
-const sameRecord = (a: Record<string, string>, b: Record<string, string>): boolean => {
-  const keys = Object.keys(a);
-  return keys.length === Object.keys(b).length && keys.every((key) => a[key] === b[key]);
-};
-
-const pairsToRecord = (pairs: Pair[]): Record<string, string> => {
-  const record: Record<string, string> = {};
-  for (const { key, value } of pairs) {
-    if (key.trim() !== "") record[key] = value;
-  }
-  return record;
-};
 
 // -- Component ----------------------------------------------------------------
 
@@ -100,5 +70,4 @@ const ExtraFieldList: React.FC<ExtraFieldListProps> = ({ values, onChange }) => 
 
 // -- Exports ------------------------------------------------------------------
 
-export { ExtraFieldList, seedPairs, reconcilePairs, sameRecord, pairsToRecord };
-export type { Pair };
+export { ExtraFieldList };
