+++ src/views/export/metadata-field-list.tsx
@@ -1,5 +1,6 @@
 import { useReconciledBuffer } from "@/hooks/useReconciledBuffer";
 import { Button } from "@/ui/button";
+import { type Row, reconcileRows, sameStrings, seedRows } from "@/views/export/metadata-field-rows";
 import { IconPlus, IconX } from "@tabler/icons-react";
 import { nanoid } from "nanoid";
 
@@ -17,23 +18,6 @@ interface MetadataFieldListProps {
   placeholder?: string;
   onChange: (next: string[]) => void;
 }
-
-interface Row {
-  id: string;
-  value: string;
-}
-
-// -- Reconciliation -----------------------------------------------------------
-
-const seedRows = (values: string[]): Row[] => values.map((value) => ({ id: nanoid(), value }));
-
-const reconcileRows = (previous: Row[], values: string[]): Row[] =>
-  values.map((value, index) =>
-    previous[index]?.value === value ? previous[index] : { id: previous[index]?.id ?? nanoid(), value },
-  );
-
-const sameStrings = (a: string[], b: string[]): boolean =>
-  a.length === b.length && a.every((value, index) => value === b[index]);
 
 // -- Component ----------------------------------------------------------------
 
@@ -92,5 +76,4 @@ const MetadataFieldList: React.FC<MetadataFieldListProps> = ({ label, itemNoun, 
 
 // -- Exports ------------------------------------------------------------------
 
-export { MetadataFieldList, INPUT_STYLES, seedRows, reconcileRows, sameStrings };
-export type { Row };
+export { MetadataFieldList, INPUT_STYLES };
